Move FirestoreCacheTestFixture into a separate source file.
